test: move tests to shared directory and update the logic
